requirements.txt
Flask==3.1.0
Flask-CORS==3.0.10
Werkzeug==3.1.3
PyYAML==6.0.2